added additional step to 'start a server'
It wasn't clear to me that the app was supposed to be loaded first (I'm new to lisp). I had to look around a little bit to find out I was missing a step.
Although this may be obvious to advanced lispers, I think it makes the documentation more beginner-friendly.